+++ clang/docs/ReleaseNotes.rst
@@ -230,6 +230,9 @@ Non-comprehensive list of changes in this release
 - Deprecated float types support from ``__builtin_elementwise_max`` and
   ``__builtin_elementwise_min``.
 
+- ``__builtin_elementwise_max`` and ``__builtin_elementwise_min`` can now be
+  used in constant expressions for floating-point arguments.
+
 - Added header ``endian.h`` which contains byte order helpers specified in POSIX
 
 - Added #pragma loop licm(disable) for llvm.loop.licm.disable metadata